+++ chaoslib/configuration.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 import os
-from typing import Dict
+from typing import Any, Dict
 
 from logzero import logger
 
@@ -10,7 +10,8 @@ from chaoslib.types import Configuration
 __all__ = ["load_configuration"]
 
 
-def load_configuration(config_info: Dict[str, str]) -> Configuration:
+def load_configuration(config_info: Dict[str, str],
+                       extra_vars: Dict[str, Any] = None) -> Configuration:
     """
     Load the configuration. The `config_info` parameter is a mapping from
     key strings to value as strings or dictionaries. In the former case, the
@@ -43,9 +44,16 @@ def load_configuration(config_info: Dict[str, str]) -> Configuration:
     variable. The `host` configuration key is dynamically fetched from the
     `HOSTNAME` environment variable, but if not defined, the default value
     `localhost` will be used instead.
+
+    When `extra_vars` is provided, it must be a dictionnary where keys map
+    to configuration key. The values from `extra_vars` always override the
+    values from the experiment itself. This is useful to the Chaos Toolkit
+    CLI mostly to allow overriding values directly from cli arguments. It's
+    seldom required otherwise.
     """
     logger.debug("Loading configuration...")
     env = os.environ
+    extra_vars = extra_vars or {}
     conf = {}
 
     for (key, value) in config_info.items():
@@ -53,12 +61,13 @@ def load_configuration(config_info: Dict[str, str]) -> Configuration:
             if value["type"] == "env":
                 env_key = value["key"]
                 env_default = value.get("default")
-                if (env_key not in env) and (env_default is None):
+                if (env_key not in env) and (env_default is None) and \
+                        (key not in extra_vars):
                     raise InvalidExperiment(
                         "Configuration makes reference to an environment key"
                         " that does not exist: {}".format(env_key))
-                conf[key] = env.get(env_key, env_default)
+                conf[key] = extra_vars.get(key, env.get(env_key, env_default))
         else:
-            conf[key] = value
+            conf[key] = extra_vars.get(key, value)
 
     return conf